requirements.txt
certifi==2020.12.5
click==7.1.2
fastapi==0.65.1
greenlet==1.1.0
h11==0.12.0
httptools==0.1.2
pika==1.2.0
pydantic==1.8.2
python-dotenv==0.17.1
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.15
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websocket-client==0.59.0
websockets==8.1
